Compilation warnings. Deprecated code. Part 1
[Issue#]       LINUXWRT-6
[Bug]          Compilation warnings
[Cause]        N/A
[Solution]     Remove them
[Verification] Build repository. Run dao tests and dpl tests. Other repositories should be able to be built: wrt-installer,wrt-plugins-common,wrt

+++ modules/widget_dao/dao/feature_dao.cpp
@@ -91,7 +91,7 @@ FeatureHandle RegisterFeature(const PluginMetafileData::Feature &feature,
 
                 WRT_DB_INSERT(insert, DeviceCapabilities, &WrtDatabase::interface())
                 insert->Values(row);
-                deviceCapID = insert->Execute();
+                deviceCapID = static_cast<int>(insert->Execute());
             }
 
             FeatureDeviceCapProxy::Row row;
